# Speedup insert and delete