… bitarray/test_bitarray.py
… test-shift_r8.py