This is related to #137 and #139.  Using shift operations (in `shift_r8()`) and the (unchanged) `copy_n()` function, but called with aligned arrays, we speedup `delete_n()`, `insert_n()`, `extend_bitarray()` and `bitarray_subscr()`.  Here are some timings on a bitarray of size 2^30:
```
 0.180 sec   insert(0, 1)  little-endian
 0.320 sec   insert(0, 1)  big-endian
 0.221 sec   pop(0)        little-endian
 0.361 sec   pop(0)        big-endian
 0.160 sec   a[1:]         little-endian
 0.301 sec   a[1:]         big-endian
 0.110 sec   a.extend      little-endian
 0.255 sec   a.extend      big-endian
```
Without this PR:
```
 8.140 sec   insert(0, 1)  little-endian
 8.010 sec   insert(0, 1)  big-endian
 6.830 sec   pop(0)        little-endian
 6.841 sec   pop(0)        big-endian
 7.990 sec   a[1:]         little-endian
 7.984 sec   a[1:]         big-endian
 8.079 sec   a.extend      little-endian
 8.081 sec   a.extend      big-endian
```